+++ packages/client/src/public-utils/hydrateObjectSetFromRid.ts
@@ -37,10 +37,12 @@ export function hydrateObjectSetFromRid<T extends ObjectOrInterfaceDefinition>(
     {
       type: "intersect",
       objectSets: [
-        {
-          type: "base",
-          objectType: definition.apiName,
-        },
+        definition.type === "interface"
+          ? { type: "interfaceBase", interfaceType: definition.apiName }
+          : {
+            type: "base",
+            objectType: definition.apiName,
+          },
         {
           type: "reference",
           reference: rid,