+++ packages/generator-converters/src/__UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition.ts
@@ -30,7 +30,12 @@ export function __UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition(
     displayName: interfaceType.displayName,
     description: interfaceType.description,
     implements: interfaceType.allExtendsInterfaces
-      ?? interfaceType.extendsInterfaces,
+      ? [...interfaceType.allExtendsInterfaces].sort((a, b) =>
+        a.localeCompare(b)
+      )
+      : interfaceType.extendsInterfaces
+      ? [...interfaceType.extendsInterfaces].sort((a, b) => a.localeCompare(b))
+      : undefined,
     properties: Object.fromEntries(
       Object.entries(interfaceType.allProperties ?? interfaceType.properties)
         .map((
@@ -44,7 +49,7 @@ export function __UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition(
               log,
             ),
           ];
-        }).filter(([key, value]) => value != null),
+        }).filter(([_, value]) => value != null),
     ),
     links: Object.fromEntries(
       Object.entries(interfaceType.allLinks ?? interfaceType.links ?? {}).map(
@@ -59,6 +64,10 @@ export function __UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition(
         }],
       ),
     ),
-    implementedBy: interfaceType.implementedByObjectTypes,
+    implementedBy: interfaceType.implementedByObjectTypes
+      ? [...interfaceType.implementedByObjectTypes].sort((a, b) =>
+        a.localeCompare(b)
+      )
+      : interfaceType.implementedByObjectTypes,
   };
 }