+++ packages/api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @osdk/api
 
+## 2.4.0-beta.11
+
+## 2.4.0-beta.10
+
+## 2.4.0-beta.9
+
+### Minor Changes
+
+- 9101bad: Adds "includeNullValues" option for exact match filters
+
+## 2.4.0-beta.8
+
+### Minor Changes
+
+- 9f4fe9e: Add new types mapping for internal creation workflow.
+
 ## 2.4.0-beta.7
 
 ## 2.4.0-beta.6