+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @osdk/client.test.ontology
 
+## 2.4.0-beta.11
+
+### Patch Changes
+
+- @osdk/api@2.4.0-beta.11
+
+## 2.4.0-beta.10
+
+### Patch Changes
+
+- @osdk/api@2.4.0-beta.10
+
+## 2.4.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [9101bad]
+  - @osdk/api@2.4.0-beta.9
+
+## 2.4.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [9f4fe9e]
+  - @osdk/api@2.4.0-beta.8
+
 ## 2.4.0-beta.7
 
 ### Patch Changes