+++ packages/client.unstable/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @osdk/client.unstable
 
+## 2.4.0-beta.11
+
+## 2.4.0-beta.10
+
+## 2.4.0-beta.9
+
+## 2.4.0-beta.8
+
 ## 2.4.0-beta.7
 
 ## 2.4.0-beta.6