+++ packages/create-app.template.tutorial-todo-aip-app.beta/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @osdk/create-app.template.tutorial-todo-aip-app.beta
 
+## 2.4.0-beta.11
+
+## 2.4.0-beta.10
+
+## 2.4.0-beta.9
+
+## 2.4.0-beta.8
+
 ## 2.4.0-beta.7
 
 ## 2.4.0-beta.6