+++ packages/generator-converters/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @osdk/generator-converters
 
+## 2.4.0-beta.11
+
+### Minor Changes
+
+- f277b30: Ensure stable ordering of arrays in code generation for object types and interface types. Sort `implements`, `implementedBy` and `linkTypes` arrays alphabetically for consistent output.
+
+### Patch Changes
+
+- @osdk/api@2.4.0-beta.11
+
+## 2.4.0-beta.10
+
+### Patch Changes
+
+- @osdk/api@2.4.0-beta.10
+
+## 2.4.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [9101bad]
+  - @osdk/api@2.4.0-beta.9
+
+## 2.4.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [9f4fe9e]
+  - @osdk/api@2.4.0-beta.8
+
 ## 2.4.0-beta.7
 
 ### Patch Changes