+++ packages/vite-plugin-oac/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @osdk/vite-plugin-oac
 
+## 0.2.0-beta.12
+
+### Patch Changes
+
+- Updated dependencies [d384b4c]
+- Updated dependencies [02be685]
+  - @osdk/maker@0.12.0-beta.14
+  - @osdk/faux@0.2.0-beta.11
+  - @osdk/cli@0.28.0-beta.12
+  - @osdk/api@2.4.0-beta.11
+  - @osdk/client.unstable@2.4.0-beta.11
+  - @osdk/generator-converters.ontologyir@2.4.0-beta.11
+
+## 0.2.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies [666787f]
+  - @osdk/maker@0.12.0-beta.12
+  - @osdk/api@2.4.0-beta.10
+  - @osdk/client.unstable@2.4.0-beta.10
+  - @osdk/generator-converters.ontologyir@2.4.0-beta.10
+  - @osdk/faux@0.2.0-beta.10
+  - @osdk/cli@0.28.0-beta.11
+
+## 0.2.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies [b0955ef]
+- Updated dependencies [defc2cf]
+- Updated dependencies [9101bad]
+  - @osdk/maker@0.12.0-beta.11
+  - @osdk/api@2.4.0-beta.9
+  - @osdk/faux@0.2.0-beta.9
+  - @osdk/cli@0.28.0-beta.10
+  - @osdk/client.unstable@2.4.0-beta.9
+  - @osdk/generator-converters.ontologyir@2.4.0-beta.9
+
+## 0.2.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [dbd13b6]
+  - @osdk/cli@0.28.0-beta.9
+
+## 0.2.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [4dffba1]
+- Updated dependencies [9f4fe9e]
+  - @osdk/maker@0.12.0-beta.10
+  - @osdk/api@2.4.0-beta.8
+  - @osdk/faux@0.2.0-beta.8
+  - @osdk/cli@0.28.0-beta.8
+  - @osdk/client.unstable@2.4.0-beta.8
+  - @osdk/generator-converters.ontologyir@2.4.0-beta.8
+
 ## 0.2.0-beta.7
 
 ### Patch Changes